requirements.txt
bibtexparser
requests
torch>=2.0.0
pyyaml
numpy
wandb
pytest